refactor: 로그인 검증 메서드 분리, 및 패키지 분리
1. Login 검증 메서드를 분리
2. Converter를 통해서 값을 반환하도록 수정
3. 패키지 구조 변경

#6
