+++ src/main/java/com/example/MaraTangOrderSystem/Converter/IngredientConverter.java
@@ -1,0 +1,14 @@
+package com.example.MaraTangOrderSystem.Converter;
+
+import com.example.MaraTangOrderSystem.dto.Ingredient.IngredientDto;
+import com.example.MaraTangOrderSystem.model.Ingredient;
+
+public class IngredientConverter {
+    public static IngredientDto convertToUserDto(Ingredient ingredient) {
+        return new IngredientDto(
+                ingredient.getId(),
+                ingredient.getName(),
+                ingredient.getPrice(),
+                ingredient.getOrderDetails());
+    }
+}
