+++ src/main/java/com/example/MaraTangOrderSystem/Converter/LoginConverter.java
@@ -1,0 +1,16 @@
+package com.example.MaraTangOrderSystem.Converter;
+
+import com.example.MaraTangOrderSystem.dto.Login.LoginResponseDto;
+import com.example.MaraTangOrderSystem.model.User;
+
+public class LoginConverter {
+    public static LoginResponseDto convertToLoginResponseDto(User user, String message) {
+        return new LoginResponseDto(
+                user.getId(),
+                user.getName(),
+                user.getNickname(),
+                user.getEmail(),
+                message
+        );
+    }
+}
